requirements.txt
PyYAML~=6.0.2
jsonschema~=4.23.0
langchain-core~=0.3.49
langchain~=0.3.21
requests~=2.32.3